requirements.txt
openai==1.3.7
keyboard==0.13.5
pyautogen==0.2.0
rope==1.11.0
radon==6.0.1
PyQt5==5.15.10
numpy==1.26.2
pandas==2.2.0
matplotlib==3.8.2
nbformat==5.9.2
libcst==1.1.0
tiktoken==0.5.2
scipy==1.11.4